NRM-445 Form clear out on failure * Form data is deleted even when submission fails * Used to check submission worked correctlyAdding a submitted_at column to db * Record submitted_at time in DB and session * Check and compare submitted_at times before deleting data https://collaboration.homeoffice.gov.uk/jira/browse/NRM-445

+++ apps/nrm/behaviours/delete-form-session.js
@@ -13,14 +13,41 @@ module.exports = superclass => class extends superclass {
         next(err);
       }
 
+      // Before deleting the report, we need to ensure that the submission time matches the one stored in the session.
+      // This prevents deletions of reports when the submission has gone wrong.
       try {
         const model = new Model();
-        const params = {
-          url: `${config.saveService.host}:${config.saveService.port}/reports/${encodeEmail(req.sessionModel.get('user-email'))}/${req.sessionModel.get('id')}`,
-          method: 'DELETE'
+        const reportUrl = `${config.saveService.host}:${config.saveService.port}/reports/${encodeEmail(req.sessionModel.get('user-email'))}/${req.sessionModel.get('id')}`;
+
+        // Get submission time from session
+        const sessionSubmissionTime = String(req.sessionModel.get('submitted-at'));
+
+        // Ger submission time from the db
+        const getParams = {
+          url: reportUrl,
+          method: 'GET'
         };
-        await model._request(params);
-        return next();
+
+        const response = await model._request(getParams);
+        const resBody = response.data;
+
+        let dbSubmissionTime = null;
+
+        if (resBody && resBody.length && resBody[0].submitted_at) {
+          dbSubmissionTime = String(resBody[0].submitted_at);
+        }
+
+        // Compare the submission times and only delete if they match
+        if(dbSubmissionTime && dbSubmissionTime === sessionSubmissionTime) {
+          const delParams = {
+            url: reportUrl,
+            method: 'DELETE'
+          };
+          await model._request(delParams);
+          return next();
+        }
+
+        throw new Error(`Submission time mismatch: ${dbSubmissionTime} does not match ${sessionSubmissionTime}`);
       } catch (error) {
         logger.error(`Error deleting data: ${error.message}`);
         return next(error);
